# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI CV Evaluator API",
    "version": "1.0.0",
    "description": "Backend service that evaluates candidate CVs and project reports asynchronously"
  },
  "servers": [{ "url": "/api" }],
  "paths": {
    "/upload": {
      "post": {
        "summary": "Upload CV and Project Report",
        "description": "Accepts a CV and project report as multipart/form-data and triggers an evaluation job.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name":{"type":"string"},
                  "cv": { "type": "string", "format": "binary" },
                  "projectReport": { "type": "string", "format": "binary" }
                },
                "required": ["cv","name","projectReport"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evaluation job created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "author": { "type": "string" },
                    "cvDocumentId": { "type": "string" },
                    "projectDocumentId": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input or file type" }
        }
      }
    },
    "/evaluate": {
      "post": {
        "summary": "Trigger evaluation using existing document IDs",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "author":{"type":"string"},
                  "jobVacancyId":{"type":"string"},
                  "cvDocumentId":{"type":"string"},
                  "projectDocumentId":{"type":"string"}
                },
                "required": [
                  "author",
                  "jobVacancyId",
                  "cvDocumentId",
                  "projectDocumentId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "status": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid payload" }
        }
      }
    },
    "/result": {
      "get": {
        "summary": "List all evaluation results",
        "description": "Retrieve a paginated list of all evaluation jobs and their statuses/results.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return (default 20).",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "default": 20 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip (default 0).",
            "required": false,
            "schema": { "type": "integer", "minimum": 0, "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of evaluation jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "status": {
                            "type": "string",
                            "enum": ["queued", "processing", "completed", "failed"]
                          },
                          "result": {
                            "type": "object",
                            "description": "Present if status is completed",
                            "properties": {
                              "cvMatchRate": { "type": "number", "example": 0.85 },
                              "cvFeedback": { "type": "string" },
                              "projectScore": { "type": "number", "example": 7.5 },
                              "projectFeedback": { "type": "string" },
                              "overallSummary": { "type": "string" }
                            }
                          }
                        }
                      }
                    },
                    "count": { "type": "integer", "example": 2 }
                  }
                }
              }
            }
          },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/result/{id}": {
      "get": {
        "summary": "Get evaluation result",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Evaluation job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status or result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "status": {
                      "type": "string",
                      "enum": ["queued", "processing", "completed", "failed"]
                    },
                    "result": {
                      "type": "object",
                      "description": "Present if status is completed",
                      "properties": {
                        "cvMatchRate": { "type": "number", "example": 0.85 },
                        "cvFeedback": { "type": "string" },
                        "projectScore": { "type": "number", "example": 7.5 },
                        "projectFeedback": { "type": "string" },
                        "overallSummary": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/job_vacancies": {
    "get": {
      "summary": "List all job vacancies",
      "description": "Returns every job vacancy document stored in the Qdrant `job_vacancies` collection.",
      "parameters": [
        {
          "name": "limit",
          "in": "query",
          "schema": { "type": "integer", "default": 1000 },
          "required": false,
          "description": "Maximum number of job records to return"
        }
      ],
      "responses": {
        "200": {
          "description": "List of job vacancies",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jobs": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          }
        },
        "500": { "description": "Server error or Qdrant failure" }
      }
    }
  }
  }
}